+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/backup_engines_list.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2026-07-01/Dpm/BackupEngines_List.json
+# x-ms-original-file: 2026-08-01/Dpm/BackupEngines_List.json
 if __name__ == "__main__":
     main()