+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/backup_resource_vault_configs_put.py
@@ -38,6 +38,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-07-01/Common/BackupResourceVaultConfigs_Put.json
+# x-ms-original-file: 2026-08-01/Common/BackupResourceVaultConfigs_Put.json
 if __name__ == "__main__":
     main()