+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/get_private_endpoint_connection_operation_status.py
@@ -39,6 +39,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-07-01/PrivateEndpointConnection/GetPrivateEndpointConnectionOperationStatus.json
+# x-ms-original-file: 2026-08-01/PrivateEndpointConnection/GetPrivateEndpointConnectionOperationStatus.json
 if __name__ == "__main__":
     main()