+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/get_tiering_cost_operation_result.py
@@ -38,6 +38,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-07-01/TieringCost/GetTieringCostOperationResult.json
+# x-ms-original-file: 2026-08-01/TieringCost/GetTieringCostOperationResult.json
 if __name__ == "__main__":
     main()