+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/protection_containers_list_with_access_type.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: 2026-07-01/AzureStorage/ProtectionContainers_List_WithAccessType.json
+# x-ms-original-file: 2026-08-01/AzureStorage/ProtectionContainers_List_WithAccessType.json
 if __name__ == "__main__":
     main()