+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/protection_policies_create_or_update_hardened.py
@@ -80,6 +80,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2026-07-01/AzureStorage/ProtectionPolicies_CreateOrUpdate_Hardened.json
+# x-ms-original-file: 2026-08-01/AzureStorage/ProtectionPolicies_CreateOrUpdate_Hardened.json
 if __name__ == "__main__":
     main()