+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/generated_samples/trigger_restore_alr_iaas_vm_restore_request_data_disk_encryption_single_entry.py
@@ -68,6 +68,6 @@ def main():
     ).result()
 
 
-# x-ms-original-file: 2026-07-01/AzureIaasVm/TriggerRestore_ALR_IaasVMRestoreRequest_DataDiskEncryption_SingleEntry.json
+# x-ms-original-file: 2026-08-01/AzureIaasVm/TriggerRestore_ALR_IaasVMRestoreRequest_DataDiskEncryption_SingleEntry.json
 if __name__ == "__main__":
     main()