+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 11.1.0 (2026-08-29)
+
+### Features Added
+
+  - Added model `InstantItemRecoveryOperationResultRequest`
+  - Model `ItemLevelRecoveryConnectionsOperations` added method `list_instant_item_recovery_operation_result`
+
 ## 11.0.0 (2026-08-13)
 
 ### Features Added